fill/requirements.txt
gdstk==0.9.59
numpy==1.26.0
pillow==11.0.0
